make int the default value, it's less annoying.
if i come across the need for int64, i will make it happen.

+++ statsdclient_test.go
@@ -8,7 +8,7 @@ import (
 
 type Experiment struct {
 	Stat     string
-	Val      int64
+	Val      int
 	Rate     float32
 	Expected string
 }
